requirements/dev.txt
-r base.txt

django-inmemorystorage==0.1.1
factory-boy==2.5.2
pytest==3.6.3
pytest-django==3.2.1
